kernel: ksud: provide ksu_handle_compat_execve_ksud v2
Since `kernel/selinux: fix pointer mismatch with 32-bit ksud on 64-bit kernels`
fixed 32-on-64 setups, it is now possible to run KernelSU properly on those.

But an issue arises around its exec hook! sys_execve passes native pointers.
But then on 32-on-64 setups, we have to use compat_sys_execve, and theres
no helper for it.

Here we provide a helper function for it, ksu_handle_compat_execve_ksud.
usage is the same as ksu_handle_execve_ksud, the only difference is that
for argv, it uses argv.ptr.compat instead of argv.ptr.native.

usage: `ksu_handle_compat_execve_ksud(filename, argv);` on compat_sys_execve

Yes this is possible to commonize and we do that on next commit.

v2 - removed struct filename dependency

Signed-off-by: backslashxx <118538522+backslashxx@users.noreply.github.com>