kernel: core_hook: migrate init_session_keyring grab to security_bprm_check
this supercedes:
	kernel: fix filp_open on older kernel's kworker (#205)
	https://github.com/tiann/KernelSU/commit/0c322a33bc22e30a9fa44a5630b9f412d601376e

No need to grab this key via LSM (key_permission) / Kprobe (key_task_permission) anymore.
ref:
- https://github.com/tiann/KernelSU/commit/afea92b829f7a5e8621b7bc33b5390106fd1377c
- https://github.com/tiann/KernelSU/commit/69d2a3ccba2476e500fdafef3b058929052c0181

as long as it passes all those checks, the keyring we can get from
it is surely init_session_keyring.

this also makes it enabled by default on 3.8~5.2

it is no longer __rcu on 5.2+ so this is just my assumption that this is
really a < 5.2 issue. its probably just because of ACK backports that it
works without it on 4.14 and 4.19 MSM/ACK.

This is just a hypothesis as hisi and samsung kernels tend to have the issue.
Well, they dont follow ACK 1:1 anyway.

checking and adding a key is not that much of an issue, since this is mostly used
on kthreads.

As for moving this to 3.8+, it on tgcred there and with that, it seems to be not needed.

This supercedes:
- kernel: expose allowlist workaround as Kconfig option
- https://github.com/backslashxx/KernelSU/commit/4c510ad8608cf2acffde87e3aaadca4dfd6a8b57

Signed-off-by: backslashxx <118538522+backslashxx@users.noreply.github.com>